refactor: 패키지 구조 변경 및 Formula 클래스 추가 1. FormulaUtil 삭제 후 Formula 객체 생성을 통해 split 2. 패키지 네이밍 변경
