+++ src/main/java/calculator/Application.java
@@ -1,14 +1,14 @@
 package calculator;
 
+import calculator.controller.FormularLauncher;
 import calculator.domain.Calculator;
-import calculator.ui.FormulaInput;
+import calculator.view.FormulaInput;
 import java.util.List;
 
 public class Application {
 
     public static void main(String[] args) {
-        List<String> formula = FormulaInput.getFormula();
-        double result = Calculator.doCalculate(formula);
-        System.out.println(result);
+        FormularLauncher formularLauncher = new FormularLauncher();
+        formularLauncher.start();
     }
 }
