+++ src/main/java/calculator/controller/FormularLauncher.java
@@ -1,0 +1,14 @@
+package calculator.controller;
+
+import calculator.domain.Calculator;
+import calculator.view.FormulaInput;
+import java.util.List;
+
+public class FormularLauncher {
+
+    public void start() {
+        List<String> formula = FormulaInput.getFormula();
+        double result = Calculator.doCalculate(formula);
+        System.out.println(result);
+    }
+}
