+++ src/main/java/calculator/domain/Formular.java
@@ -1,0 +1,12 @@
+package calculator.domain;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
+public class Formular {
+
+    public static List<String> stringSpliter(String expressions) {
+        return new ArrayList<>(Arrays.asList(expressions.split(" ")));
+    }
+}
