+++ src/main/java/calculator/view/FormulaOutput.java
@@ -1,4 +1,4 @@
-package calculator.ui;
+package calculator.view;
 
 public class FormulaOutput {
