+++ src/main/java/racingcar/controller/GameLauncher.java
@@ -8,4 +8,3 @@ public class GameLauncher {
     }
 
 }
-
